Revert "NetCDF with 1D lat-lon cannot be parsed into fieldlist (#658)"
This reverts commit 4ed4526a6fe29b570754965d013b2c81e78b0315.
