+++ src/earthkit/data/readers/netcdf/dataset.py
@@ -41,8 +41,8 @@ class DataSet:
     def bbox(self, variable):
         data_array = self[variable]
 
-        keys, dims = self._get_xy_dims(data_array)
-        key = ("bbox", tuple(keys), tuple(dims))
+        keys, coords = self._get_xy_coords(data_array)
+        key = ("bbox", tuple(keys), tuple(coords))
         if key in self._cache:
             return self._cache[key]
 
@@ -59,10 +59,7 @@ class DataSet:
         self._cache[key] = (north, west, south, east)
         return self._cache[key]
 
-    def _get_xy_dims(self, data_array):
-        # TODO: refactor it
-
-        # the last 2 dimensions are y,x
+    def _get_xy_coords(self, data_array):
         if (
             len(data_array.dims) >= 2
             and data_array.dims[-1] in GEOGRAPHIC_COORDS["x"]
@@ -70,68 +67,40 @@ class DataSet:
         ):
             return ("y", "x"), (data_array.dims[-2], data_array.dims[-1])
 
-        # the last 2 dimensions are x, y
-        if (
-            len(data_array.dims) >= 2
-            and data_array.dims[-1] in GEOGRAPHIC_COORDS["y"]
-            and data_array.dims[-2] in GEOGRAPHIC_COORDS["x"]
-        ):
-            return ("x", "y"), (data_array.dims[-2], data_array.dims[-1])
-
-        # try to find the x and y dimensions
         keys = []
-        dims = []
+        coords = []
         axes = ("x", "y")
         for dim in data_array.dims:
             for ax in axes:
                 candidates = GEOGRAPHIC_COORDS.get(ax, [])
                 if dim in candidates:
                     keys.append(ax)
-                    dims.append(dim)
+                    coords.append(dim)
                 else:
                     ax = data_array.coords[dim].attrs.get("axis", "").lower()
                     if ax in axes:
                         keys.append(ax)
-                        dims.append(dim)
+                        coords.append(dim)
             if len(keys) == 2:
-                return tuple(keys), tuple(dims)
-
-        # 1D geographic coordinates using the dim='values' convention
-        if not keys or not dims:
-            dim = "values"
-            if data_array.dims and data_array.dims[-1] == dim:
-                for x, y in zip(GEOGRAPHIC_COORDS["x"], GEOGRAPHIC_COORDS["y"]):
-                    if (x in data_array.coords or x in self._ds.variables) and (
-                        y in data_array.coords or y in self._ds.variables
-                    ):
-                        if (
-                            len(self._ds[x].dims) == 1
-                            and len(self._ds[y].dims) == 1
-                            and self._ds[x].dims[-1] == dim
-                            and self._ds[y].dims[-1] == dim
-                        ):
-                            return tuple(["x"]), tuple([dim])
+                return tuple(keys), tuple(coords)
 
         for ax in axes:
             if ax not in keys:
-                raise ValueError(f"No dimension found with axis '{ax}'")
+                raise ValueError(f"No coordinate found with axis '{ax}'")
 
-        return keys, dims
+        return keys, coords
 
     def _get_xy(self, data_array, flatten=False, dtype=None):
-        # TODO: refactor it
-
-        keys, dims = self._get_xy_dims(data_array)
-        key = ("grid_points", tuple(keys), tuple(dims))
+        keys, coords = self._get_xy_coords(data_array)
+        key = ("grid_points", tuple(keys), tuple(coords))
 
         if key in self._cache:
             points = self._cache[key]
         else:
             points = dict()
-            if all(d in data_array.coords for d in dims):
-                v0, v1 = data_array.coords[dims[0]], data_array.coords[dims[1]]
-                points[keys[1]], points[keys[0]] = np.meshgrid(v1, v0)
-                self._cache[key] = points
+            v0, v1 = data_array.coords[coords[0]], data_array.coords[coords[1]]
+            points[keys[1]], points[keys[0]] = np.meshgrid(v1, v0)
+            self._cache[key] = points
 
         if flatten:
             points["x"] = points["x"].reshape(-1)
@@ -143,8 +112,7 @@ class DataSet:
             return points["x"], points["y"]
 
     def _get_latlon(self, data_array, flatten=False, dtype=None):
-        # TODO: refactor it
-        keys, dims = self._get_xy_dims(data_array)
+        keys, coords = self._get_xy_coords(data_array)
 
         points = dict()
 
@@ -158,27 +126,22 @@ class DataSet:
         latitude = _get_ll(lat_keys)
         longitude = _get_ll(lon_keys)
 
-        # lat-lon variable/coordinate available
         if latitude is not None and longitude is not None:
-            if latitude.dims == dims and longitude.dims == dims:
+            if latitude.dims == coords and longitude.dims == coords:
                 latitude = latitude.data
                 longitude = longitude.data
                 points["y"] = latitude
                 points["x"] = longitude
 
-        # lat-lon meshgrid
         if not points:
-            key = ("grid_points", tuple(keys), tuple(dims))
+            key = ("grid_points", tuple(keys), tuple(coords))
 
             if key in self._cache:
                 points = self._cache[key]
             else:
-                if all(d in data_array.coords for d in dims):
-                    v0, v1 = data_array.coords[dims[0]], data_array.coords[dims[1]]
-                    points[keys[1]], points[keys[0]] = np.meshgrid(v1, v0)
-                    self._cache[key] = points
-                else:
-                    raise ValueError(f"Could not find lat-lon coordinates for {data_array.name}")
+                v0, v1 = data_array.coords[coords[0]], data_array.coords[coords[1]]
+                points[keys[1]], points[keys[0]] = np.meshgrid(v1, v0)
+                self._cache[key] = points
 
         if flatten:
             points["x"] = points["x"].reshape(-1)
